+++ renderer/src/ui/divider.jsx
@@ -1,0 +1,3 @@
+import {React} from "modules";
+
+export default ({className}) => <div className={`bd-divider ${className || ""}`}></div>;